create probe module
