+++ src/probe.rs
@@ -1,0 +1,93 @@
+use std::str::FromStr;
+
+use anyhow::{anyhow, bail};
+use probe_rs::{DebugProbeInfo, Probe};
+
+use crate::cli;
+
+pub(crate) fn open(opts: &cli::Opts) -> Result<Probe, anyhow::Error> {
+    let all_probes = Probe::list_all();
+    let filtered_probes = if let Some(probe_opt) = opts.probe.as_deref() {
+        let selector = probe_opt.parse()?;
+        filter(&all_probes, &selector)
+    } else {
+        all_probes
+    };
+
+    if filtered_probes.is_empty() {
+        bail!("no probe was found")
+    }
+
+    log::debug!("found {} probes", filtered_probes.len());
+
+    if filtered_probes.len() > 1 {
+        let _ = print(filtered_probes);
+        bail!("more than one probe found; use --probe to specify which one to use");
+    }
+
+    let probe = filtered_probes[0].open()?;
+    log::debug!("opened probe");
+    Ok(probe)
+}
+
+pub(crate) fn print(probes: Vec<DebugProbeInfo>) {
+    if !probes.is_empty() {
+        println!("The following devices were found:");
+        probes
+            .iter()
+            .enumerate()
+            .for_each(|(num, link)| println!("[{}]: {:?}", num, link));
+    } else {
+        println!("No devices were found.");
+    }
+}
+
+fn filter(probes: &[DebugProbeInfo], selector: &ProbeFilter) -> Vec<DebugProbeInfo> {
+    probes
+        .iter()
+        .filter(|&p| {
+            if let Some((vid, pid)) = selector.vid_pid {
+                if p.vendor_id != vid || p.product_id != pid {
+                    return false;
+                }
+            }
+
+            if let Some(serial) = &selector.serial {
+                if p.serial_number.as_deref() != Some(serial) {
+                    return false;
+                }
+            }
+
+            true
+        })
+        .cloned()
+        .collect()
+}
+
+struct ProbeFilter {
+    vid_pid: Option<(u16, u16)>,
+    serial: Option<String>,
+}
+
+impl FromStr for ProbeFilter {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let parts = s.split(':').collect::<Vec<_>>();
+        match &*parts {
+            [serial] => Ok(Self {
+                vid_pid: None,
+                serial: Some(serial.to_string()),
+            }),
+            [vid, pid] => Ok(Self {
+                vid_pid: Some((u16::from_str_radix(vid, 16)?, u16::from_str_radix(pid, 16)?)),
+                serial: None,
+            }),
+            [vid, pid, serial] => Ok(Self {
+                vid_pid: Some((u16::from_str_radix(vid, 16)?, u16::from_str_radix(pid, 16)?)),
+                serial: Some(serial.to_string()),
+            }),
+            _ => Err(anyhow!("invalid probe filter")),
+        }
+    }
+}
